backend/requirements.txt
# Core FastAPI
fastapi==0.129.0
uvicorn[standard]>=0.31.0
gunicorn>=23.0.0

# Database
sqlalchemy>=2.0.37
psycopg2-binary>=2.9.11
pymongo>=4.11.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Security & Crypto
passlib[bcrypt]==1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.22
pycryptodome>=3.21.0
cryptography>=44.0.0

# Utilities
python-dotenv>=1.0.1
psutil>=6.1.0
python-dateutil>=2.9.0
httpx>=0.28.1

# Web3 & Payments
setuptools>=75.8.0
stripe>=11.4.0
web3>=7.7.0
eth-account>=0.13.4

# OAuth & JWT
authlib>=1.5.0
python-jose[cryptography]>=3.4.0
PyJWT>=2.10.0

# Email & Validation
aiosmtplib>=3.0.2
email-validator>=2.2.0

# Additional
redis>=5.3.0
resend>=2.4.0